+++ client/src/components/NavBar/DropMenu/DropMenu.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import '../../App.css';
 import Menu from './Menu'
 
 const DropDown = (props) => {
