+++ client/src/components/NavBar/DropMenu/Menu.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import '../../App.css'
 
 const Menu = (props) => (
