+++ client/src/components/NavBar/SearchBar.js
